Move custom class into dashboard
