+++ dashboard/templates/stable/stable1.html
@@ -1,3 +1,6 @@
+<!--{% extends layout %}-->
+<!--{% block page_body %}-->
+<!--{% include "components/alert_error.html" %}-->
 <div>
     <table>
         <tr>
@@ -16,3 +19,4 @@
         {% endif %}
     </table>
 </div>
+<!--{% endblock %}-->
