+++ dashboard/templates/stable/text_annotation_example.html
